added user Services and Repositories

+++ src/main/java/com/petproject/minivns/repositories/UserRepository.java
@@ -4,5 +4,6 @@ package com.petproject.minivns.repositories;
 import com.petproject.minivns.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
 }
